# Optimise ECAL multifit (for cmssw 92x)