--- a/RecoLocalCalo/EcalRecProducers/plugins/EcalUncalibRecHitProducer.cc
+++ b/RecoLocalCalo/EcalRecProducers/plugins/EcalUncalibRecHitProducer.cc
@@ -116,21 +116,11 @@ EcalUncalibRecHitProducer::produce(edm::Event& evt, const edm::EventSetup& es) {
 
         // loop over EB digis
         if (ebDigis)
-        {
-                ebUncalibRechits->reserve(ebDigis->size());
-                for(EBDigiCollection::const_iterator itdg = ebDigis->begin(); itdg != ebDigis->end(); ++itdg) {
-                        worker_->run(evt, itdg, *ebUncalibRechits);
-                }
-        }
+            worker_->run(evt, *ebDigis, *ebUncalibRechits);
 
         // loop over EB digis
         if (eeDigis)
-        {
-                eeUncalibRechits->reserve(eeDigis->size());
-                for(EEDigiCollection::const_iterator itdg = eeDigis->begin(); itdg != eeDigis->end(); ++itdg) {
-                        worker_->run(evt, itdg, *eeUncalibRechits);
-                }
-        }
+            worker_->run(evt, *eeDigis, *eeUncalibRechits);
 
         // put the collection of recunstructed hits in the event
         evt.put(std::move(ebUncalibRechits), ebHitCollection_);