loop over DIGIs inside EcalUncalibRecHitWorker*::run()
compile EcalRecProducers and EcalRecAlgos with `-O3`

from @felicepantaleo :
  - faster check for a positive vector
  - increase threshold with a smoother granularity

from @VinInn :
  - use lazyProduct() and noalias() in NNLS