@fwyzard 2017-05-07
In principle we do not expect any change in the physics results, while we have observed a speed up of 12-13% running the full multifit at HLT.