@cmsbuild 2017-05-07
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18600/19623/summary.html

Comparison Summary:
 - No significant changes to the logs found
 - Reco comparison results: 1692 differences found in the comparisons
 - DQMHistoTests: Total files compared: 24
 - DQMHistoTests: Total histograms compared: 1831084
 - DQMHistoTests: Total failures: 11683
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 1819221
 - DQMHistoTests: Total skipped: 180
 - DQMHistoTests: Total Missing objects: 0
 - Checked 98 log files, 14 edm output root files, 24 DQM output files