@VinInn 2017-05-08
At best of my knowledge the default is an augmented O2 not O3.
not the first time Ofast is slower in particular with c++, in particular with gcc6.3, in particular in presence of short loops.... (search gcc bugzilla: essentially overhead overwhelms gain)
anyhow eigen uses its own "vectorized intrinsics": so little to be gain from the compiler here...