@slava77 2017-05-08
@fwyzard 
How much faster is the result with default 02/02 compared to this new proposed O3/O3?
I'm fine dropping Ofast, especially given the intel-amd reproducibility issue.