@fwyzard 2017-05-08
Hi Slava,
I have summarised the impact of all the changes here: https://patatrack.web.cern.ch/content/ecal-multifit-hlt-optimization .