@slava77 2017-05-10
On 5/8/17 10:28 AM, Andrea Bocci wrote:
> Hi Slava,
> I have summarised the impact of all the changes here:
> https://patatrack.web.cern.ch/content/ecal-multifit-hlt-optimization .
> 

Hi Andrea,


From the preceding discussion, I was hoping that message meant that a
check with default flags was done.
I do not see it in the table.

Please check by removing the "flags" in xml to use the default in both
cases.
1% or smaller (but still stat significant) difference between that and
the -O3 would be too small to argue for a non-standard setting.



> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-sw/cmssw/pull/18600#issuecomment-299933655>, or
> mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEdcbgkMsf2seSsfYu1_ITLQnkiLW51wks5r31C0gaJpZM4NS9j0>.
>