@fwyzard 2017-05-10
> Please check by removing the "flags" in xml to use the default in both cases.

Sorry, but I have tested the flags we are currently using and those we propose, and the change shows a clear gain. Trying something else is a valid suggestion, but should not be an objection to these changes until somebody shows we can do better.

> 1% or smaller (but still stat significant) difference between that and the -O3 would be too small to argue for a non-standard setting.

I might agree if not for the fact that we are already sing a non-standard setting.
However the HLT *is* cpu-limted, and this module is the single most cpu-intensive one run at HLT, so I'll take a 1% improvement if it is clearly reproducible.

On the other hand it would be nice if the ECAL people would comment on the changes, and made a more through check of the (absence of) impact on the physics results.