@slava77 2017-05-10
On 5/10/17 10:28 AM, Andrea Bocci wrote:
>     Please check by removing the "flags" in xml to use the default in
>     both cases.
> 
> Sorry, but I have tested the flags we are currently using and those we
> propose, and the change shows a clear gain. Trying something else is a
> valid suggestion, but should not be an objection to these changes until
> somebody shows we can do better.


Use of non-standard flags is an exception.
To change them I'm asking for a justification that this is better than
standard default flags.
An argument that it's better than a current exception is not good enough
for me.



> 
>     1% or smaller (but still stat significant) difference between that
>     and the -O3 would be too small to argue for a non-standard setting.
> 
> I might agree if not for the fact that we are already sing a
> non-standard setting.
> However the HLT /is/ cpu-limted, and this module is the single most
> cpu-intensive one run at HLT, so I'll take a 1% improvement if it is
> clearly reproducible.
> 
> On the other hand it would be nice if the ECAL people would comment on
> the changes, and made a more through check of the (absence of) impact on
> the physics results.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-sw/cmssw/pull/18600#issuecomment-300554609>, or
> mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEdcbrxL6OkOK7lWtEpo5NnesHbXFK6Mks5r4fOggaJpZM4NS9j0>.
>