@slava77 2017-05-10
on 

> On the other hand it would be nice if the ECAL people would comment on the changes, and made a 
> more through check of the (absence of) impact on the physics results.

Was it presented to ECAL DPG already?
@argiro please check and confirm that changes are OK for you.