@fwyzard 2017-05-10
So, maybe I was not expecting "thanks for optimising this piece of *offline reconstruction* code"... but I was definitely *not* expecting "the improvement is not large enough to justify making a change".