@slava77 2017-05-10
On 5/10/17 10:56 AM, Andrea Bocci wrote:
> So, maybe I was not expecting "thanks for optimising this piece of
> /offline reconstruction/ code"... but I was definitely /not/ expecting
> "the improvement is not large enough to justify making a change".

I think we are discussing the flag use, not the remaining valuable work
on code updates in this PR.


> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-sw/cmssw/pull/18600#issuecomment-300563510>, or
> mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEdcbuQJJ_rilKT3q6h5d9qi_FtOKSrNks5r4fpYgaJpZM4NS9j0>.
>