@slava77 2017-05-10
@fwyzard 
I see no changes in DQM and fwlite-based reco comparisons based on jenkins tests, except for numerical-level changes that apparently start from ECAL and are visible only in DQM histograms (apparently most of the are fit-based, more susceptible to floating point precision ).
This only covers somewhat short tests (up to 100 events, with somewhat moderate pileup in data wflows).

Did you have in mind there could be other places or setups where changes can be expected?