@argiro 2017-05-11
the Ecal people, in particular the authors of the multifit, took a quick look without gaining a deep understanding of the optimization under discussion (no opinion on the compiler flag in particular).
In a pragmatic approach, although, as per human nature, a bit suspicious when "alien" ( ;-) ) hands touch their code, they trust the 
comparison tests and the reputation of the fiddler , and are happy with the gain in speed, thank you !

> On 10 May 2017, at 19:49, Slava Krutelyov <notifications@github.com> wrote:
> 
> on
> 
> On the other hand it would be nice if the ECAL people would comment on the changes, and made a
> more through check of the (absence of) impact on the physics results.
> 
> Was it presented to ECAL DPG already?
> @argiro please check and confirm that changes are OK for you.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.
>