@fwyzard 2017-05-11
Back about compiler flags:

> An argument that it's better than a current exception is not good enough for me.

I do stand by my argument: 1% better *is* good enough for me.
Fortunately looks like there is a solution that we both like better:

| code base  | cpu time | gain | gain |
|----|---|---|---|
| CMSSW 9.1.0-pre3               | 34.0 ± 0.1 ms| reference 
| all code changes, -O2 / -Ofast | 30.5 ± 0.1 ms|    -10.5% | reference
| all code changes, -O3          | 30.0 ± 0.6 ms|    -11.9% |     -1.5%
| all code changes, -O2          | 29.2 ± 0.1 ms|    -14.3% |     -4.2%

So, plain `-O2` is faster than both `-Ofast` and `-O3`